minor refacto for schematics - feature schematic also add import to own module to parent module
