+++ schematics/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './file-manipulation';
+export * from './index-manipulation';
+export * from './module-manipulation';
